Remove cloud controller manager from metadata package
